requirements.txt
anyio==4.11.0
Authlib==1.6.5
authutils==7.1.0
bcrypt==5.0.0
blinker==1.9.0
cached-property==1.5.2
cdiserrors==1.0.0
cdislogging==1.1.1
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
django-timedeltafield==0.7.10
dnspython==2.8.0
dominate==2.9.1
dotenv==0.9.9
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
h11==0.16.0
hashids==1.3.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jwt==1.4.0
MarkupSafe==3.0.2
mongoengine==0.29.1
pillow==12.0.0
pycparser==2.23
PyJWT==2.10.1
pymongo==4.15.3
python-dotenv==1.1.1
requests==2.32.3
sniffio==1.3.1
urllib3==2.3.0
visitor==0.1.3
Werkzeug==3.1.3
xmltodict==0.15.1
